Switch to `bstr` & rework `WordTally` impl

+++ src/input.rs
@@ -53,9 +53,11 @@ impl Input {
 
     /// Get the size of the input in bytes, if available.
     /// Returns None for stdin or if the file size can't be determined.
-    pub fn size(&self) -> Option<u64> {
+    pub fn size(&self) -> Option<usize> {
         match self {
-            Self::File(path) => fs::metadata(path).map(|metadata| metadata.len()).ok(),
+            Self::File(path) => fs::metadata(path)
+                .map(|metadata| metadata.len() as usize)
+                .ok(),
             Self::Stdin => None,
         }
     }
